# openapi.json
{"openapi":"3.1.0","info":{"title":"Company Analysis API","description":"MVP API for company intelligence analysis using Gemini AI","version":"1.0.0"},"paths":{"/auth/token":{"post":{"tags":["authentication"],"summary":"Get Access Token","description":"Generate access token for API authentication","operationId":"get_access_token_auth_token_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/gemini-key":{"put":{"tags":["admin"],"summary":"Update Gemini Key","description":"Update Gemini API key","operationId":"update_gemini_key_admin_gemini_key_put","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGeminiKeyRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGeminiKeyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/companies/search":{"post":{"tags":["companies"],"summary":"Search Company Endpoint","description":"Search for company analysis with fuzzy matching and auto-generation","operationId":"search_company_endpoint_companies_search_post","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"type":"string","title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanySearchRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/CompanySearchResponse"},{"$ref":"#/components/schemas/CompanyNotFoundResponse"}],"title":"Response Search Company Endpoint Companies Search Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/companies/{company_id}":{"get":{"tags":["companies"],"summary":"Get Company Analysis","description":"Get specific company analysis by ID","operationId":"get_company_analysis_companies__company_id__get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"integer","title":"Company Id"}},{"name":"authorization","in":"header","required":true,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanySearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CompanyNotFoundResponse":{"properties":{"error":{"type":"string","title":"Error"},"message":{"type":"string","title":"Message"},"suggestions":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Suggestions"}},"type":"object","required":["error","message"],"title":"CompanyNotFoundResponse"},"CompanySearchRequest":{"properties":{"company_name":{"type":"string","title":"Company Name"}},"type":"object","required":["company_name"],"title":"CompanySearchRequest"},"CompanySearchResponse":{"properties":{"id":{"type":"integer","title":"Id"},"company_name":{"type":"string","title":"Company Name"},"canonical_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Canonical Name"},"analysis_result":{"type":"object","title":"Analysis Result"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","company_name","canonical_name","analysis_result","status","created_at"],"title":"CompanySearchResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TokenRequest":{"properties":{"client_id":{"type":"string","title":"Client Id"},"client_secret":{"type":"string","title":"Client Secret"}},"type":"object","required":["client_id","client_secret"],"title":"TokenRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"type":"integer","title":"Expires In"}},"type":"object","required":["access_token","expires_in"],"title":"TokenResponse"},"UpdateGeminiKeyRequest":{"properties":{"new_gemini_api_key":{"type":"string","title":"New Gemini Api Key"}},"type":"object","required":["new_gemini_api_key"],"title":"UpdateGeminiKeyRequest"},"UpdateGeminiKeyResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"UpdateGeminiKeyResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}